requirements.txt
datasets>=2.19.0
requests>=2.31.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.1
tqdm>=4.66.0
